Move NewCard back to top level and add an unimplemented card helper method
