+++ server/game/core/card/propertyMixins/Token.ts
@@ -3,7 +3,7 @@ import { IActionAbilityProps, IEpicActionProps } from '../../../Interfaces';
 import { CardActionAbility } from '../../ability/CardActionAbility';
 import Card from '../Card';
 import { InPlayCardConstructor } from '../baseClasses/InPlayCard';
-import { CardConstructor } from '../baseClasses/NewCard';
+import { CardConstructor } from '../NewCard';
 
 // TODO TOKENS: custom defeat logic and any other logic
 export function Token<TBaseClass extends InPlayCardConstructor>(BaseClass: TBaseClass) {
